Split demo into files
